+++ queues/qclient_cache.py
@@ -1,7 +1,7 @@
 import threading
 
-from .rabbit_queue import RMQueue
-from .util import check_for_kb_interrupt
+from queues.rabbit_queue import RMQueue
+from util.util import check_for_kb_interrupt
 
 _queue_cache = {}
