+++ studio/firebase_provider.py
@@ -1,4 +1,4 @@
-from .keyvalue_provider import KeyValueProvider
+from db_providers.keyvalue_provider import KeyValueProvider
 from .firebase_storage_handler import FirebaseStorageHandler
 
 class FirebaseProvider(KeyValueProvider):
